chore(workflows): move the strip-placeholder command tests to their own file
